+++ CHANGELOG.md
@@ -45,6 +45,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/107]:** com.google.fonts/check/metadata/normal_style
   - **[com.google.fonts/check/108]:** com.google.fonts/check/metadata/nameid/family_and_full_names
   - **[com.google.fonts/check/109]:** com.google.fonts/check/metadata/fontname_not_camel_cased
+  - **[com.google.fonts/check/166]:** com.google.fonts/check/fontv
 
 ### Other relevant code-changes
   - We temporarily disabled com.google.fonts/check/metadata/match_filename_postscript for variable fonts until we have a clear definition of the VF naming rules as discussed at https://github.com/google/fonts/issues/1817